requirements.txt
astroid==2.5.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.3
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.26
protobuf==3.17.3
pycodestyle==2.7.0
pylint==2.8.2
python-dotenv==1.1.1
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
